pyproject.toml
[tool.poetry]
name = "kaleriavvs0611"
version = "0.1.0"
description = ""
authors = ["VasilevAlexandr97 <sashaorpix@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
vkbottle = "^4.3.2"
uvloop = "^0.16.0"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.0"}
aerich = "^0.6.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"


[tool.aerich]
tortoise_orm = "vk_bot.database.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
